fastapi-server/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
wikipedia-api
tqdm
nltk